refactor(api): move config and logger types to types package
switch config routes over to new method
