--- packages/api/src/api/config/config.routes.ts
@@ -1,17 +1,0 @@
-import { ServerRoute } from "hapi";
-
-import { GET, PUT } from "../../util/hapiRoute";
-import { getAppConfig, putUpdateConfig } from "./config.handler";
-
-export const routes: ServerRoute[] = [
-  {
-    method: GET,
-    path: "/config",
-    handler: getAppConfig
-  },
-  {
-    method: PUT,
-    path: "/config",
-    handler: putUpdateConfig
-  }
-];
