--- packages/api/src/api/config/index.ts
@@ -1,1 +1,0 @@
-export { routes as config } from "./config.routes";
