+++ packages/api/src/config/index.ts
@@ -1,2 +1,1 @@
-export * from "./schema";
 export * from "./config";
