+++ packages/endpoints/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./route";
-
+export * from "./api/config";
 export * from "./api/games";
