+++ packages/types/package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "@nxbm/tslint-rules": "^0.0.0",
+    "convict": "^4.3.2",
     "rimraf": "^2.6.1",
     "tslint": "^5.8.0",
     "typescript": "^3.0.3"
