+++ packages/types/src/index.ts
@@ -1,1 +1,4 @@
 export * from "./files";
+export * from "./config";
+export * from "./logger";
+export * from "./schema";
